fixed to store/restore TimelineAdapter
