+++ app/src/androidTest/java/com/freshdigitable/udonroad/MainToUserInfoActivityInstTest.java
@@ -16,9 +16,13 @@
 
 package com.freshdigitable.udonroad;
 
+import android.app.Activity;
 import android.support.test.espresso.Espresso;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
+import android.support.test.runner.lifecycle.ActivityLifecycleMonitorRegistry;
+import android.support.test.runner.lifecycle.Stage;
+import android.support.v4.app.Fragment;
 import android.support.v7.widget.Toolbar;
 
 import com.freshdigitable.udonroad.util.PerformUtil;
@@ -26,6 +30,9 @@ import com.freshdigitable.udonroad.util.PerformUtil;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
+import java.util.Collection;
+import java.util.List;
 
 import twitter4j.Relationship;
 import twitter4j.ResponseList;
@@ -40,9 +47,14 @@ import static android.support.test.espresso.matcher.ViewMatchers.isAssignableFro
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static com.freshdigitable.udonroad.util.AssertionUtil.anywayNotVisible;
 import static com.freshdigitable.udonroad.util.AssertionUtil.checkMainActivityTitle;
+import static com.freshdigitable.udonroad.util.IdlingResourceUtil.getSimpleIdlingResource;
+import static com.freshdigitable.udonroad.util.IdlingResourceUtil.runWithIdlingResource;
+import static com.freshdigitable.udonroad.util.MatcherUtil.onIFFAB;
 import static com.freshdigitable.udonroad.util.StatusViewMatcher.ofStatusView;
 import static com.freshdigitable.udonroad.util.StatusViewMatcher.ofStatusViewAt;
+import static com.freshdigitable.udonroad.util.TwitterResponseMock.createStatus;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -78,6 +90,7 @@ public class MainToUserInfoActivityInstTest extends TimelineInstTestBase {
     // tear down
     Espresso.pressBack();
     checkMainActivityTitle(R.string.title_home);
+    onIFFAB().check(anywayNotVisible());
   }
 
   @Test
@@ -85,15 +98,20 @@ public class MainToUserInfoActivityInstTest extends TimelineInstTestBase {
     PerformUtil.clickUserIconAt(0);
     onView(withId(R.id.user_screen_name)).check(matches(screenNameMatcher));
     Espresso.pressBack();
+    onIFFAB().check(anywayNotVisible());
+
     PerformUtil.clickUserIconAt(0);
     onView(withId(R.id.user_screen_name)).check(matches(screenNameMatcher));
     Espresso.pressBack();
+    onIFFAB().check(anywayNotVisible());
+
     PerformUtil.clickUserIconAt(0);
     onView(withId(R.id.user_screen_name)).check(matches(screenNameMatcher));
 
     // tear down
     Espresso.pressBack();
     checkMainActivityTitle(R.string.title_home);
+    onIFFAB().check(anywayNotVisible());
   }
 
   @Test
@@ -107,6 +125,7 @@ public class MainToUserInfoActivityInstTest extends TimelineInstTestBase {
 
     onView(ofStatusViewAt(R.id.timeline, 0))
         .check(matches(ofStatusView(withText(findByStatusId(20000).getText()))));
+    onIFFAB().check(anywayNotVisible());
   }
 
   @Test
@@ -117,6 +136,50 @@ public class MainToUserInfoActivityInstTest extends TimelineInstTestBase {
 
     Espresso.pressBack();
     checkMainActivityTitle(R.string.title_home);
+  }
+
+  @Test
+  public void receiveStatusWhenUserInfoIsAppeared_then_timelineIsNotScrolled() throws Exception {
+    final Status top = findByStatusId(20000);
+    onView(ofStatusViewAt(R.id.timeline, 0)).check(matches(ofStatusView(withText(top.getText()))));
+
+    PerformUtil.clickUserIconAt(0);
+    runWithIdlingResource(getSimpleIdlingResource("userInfo", () -> {
+      final Collection<Activity> activities = ActivityLifecycleMonitorRegistry.getInstance().getActivitiesInStage(Stage.RESUMED);
+      for (Activity a : activities) {
+        if (!(a instanceof UserInfoActivity)) {
+          continue;
+        }
+        final List<Fragment> fragments = ((UserInfoActivity) a).getSupportFragmentManager().getFragments();
+        for (Fragment f : fragments) {
+          if (f instanceof UserInfoPagerFragment) {
+            return true;
+          }
+        }
+      }
+      return false;
+    }), () ->
+        onView(withText("TWEET\n20")).check(matches(isDisplayed())));
+
+    final Status received22 = createStatus(22000);
+    receiveStatuses(false, received22);
+    Espresso.pressBack();
+    runWithIdlingResource(getSimpleIdlingResource("back to main", () ->
+        getTimelineView().getAdapter().getItemCount() == 21), () ->
+        onView(ofStatusViewAt(R.id.timeline, 0)).check(matches(ofStatusView(withText(top.getText())))));
+    onIFFAB().check(anywayNotVisible());
+
+    final Status received25 = createStatus(25000);
+    receiveStatuses(received25);
+    onView(ofStatusViewAt(R.id.timeline, 0)).check(matches(ofStatusView(withText(top.getText()))));
+
+    PerformUtil.clickHeadingOnMenu();
+    onView(ofStatusViewAt(R.id.timeline, 0))
+        .check(matches(ofStatusView(withText(received25.getText()))));
+    onView(ofStatusViewAt(R.id.timeline, 1))
+        .check(matches(ofStatusView(withText(received22.getText()))));
+    onView(ofStatusViewAt(R.id.timeline, 2))
+        .check(matches(ofStatusView(withText(top.getText()))));
   }
 
   @Override
